Restructure into workspaces
