+++ cramjam-core/src/exceptions.rs
