+++ src/brotli.rs
@@ -1,7 +1,7 @@
 //! brotli de/compression interface
-use crate::exceptions::{CompressionError, DecompressionError};
-use crate::io::RustyBuffer;
 use crate::{to_py_err, BytesType};
+use cramjam_core::exceptions::{CompressionError, DecompressionError};
+use cramjam_core::io::RustyBuffer;
 use pyo3::prelude::*;
 use pyo3::wrap_pyfunction;
 use pyo3::PyResult;
